TASK: Doctrine migration: Rename index for better performance
Adjusts the doctrine migration to use RENAME INDEX when available
(with MySQL version 5.7+)

+++ Neos.Flow/Migrations/Mysql/Version20170110130149.php
@@ -2,6 +2,7 @@
 namespace Neos\Flow\Persistence\Doctrine\Migrations;
 
 use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\DBAL\Platforms\MySQL57Platform;
 use Doctrine\DBAL\Schema\Schema;
 
 class Version20170110130149 extends AbstractMigration
@@ -23,8 +24,13 @@ class Version20170110130149 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on "mysql".');
 
-        $this->addSql('DROP INDEX flow_identity_typo3_flow_security_account ON neos_flow_security_account');
-        $this->addSql('CREATE UNIQUE INDEX flow_identity_neos_flow_security_account ON neos_flow_security_account (accountidentifier, authenticationprovidername)');
+        // Renaming of indexes is only possible with MySQL version 5.7+
+        if ($this->connection->getDatabasePlatform() instanceof MySQL57Platform) {
+            $this->addSql('ALTER TABLE neos_flow_security_account RENAME INDEX flow_identity_typo3_flow_security_account TO flow_identity_neos_flow_security_account');
+        } else {
+            $this->addSql('DROP INDEX flow_identity_typo3_flow_security_account ON neos_flow_security_account');
+            $this->addSql('CREATE UNIQUE INDEX flow_identity_neos_flow_security_account ON neos_flow_security_account (accountidentifier, authenticationprovidername)');
+        }
     }
 
     /**
@@ -35,7 +41,12 @@ class Version20170110130149 extends AbstractMigration
     {
         $this->abortIf($this->connection->getDatabasePlatform()->getName() != 'mysql', 'Migration can only be executed safely on "mysql".');
 
-        $this->addSql('DROP INDEX flow_identity_neos_flow_security_account ON neos_flow_security_account');
-        $this->addSql('CREATE UNIQUE INDEX flow_identity_typo3_flow_security_account ON neos_flow_security_account (accountidentifier, authenticationprovidername)');
+        // Renaming of indexes is only possible with MySQL version 5.7+
+        if ($this->connection->getDatabasePlatform() instanceof MySQL57Platform) {
+            $this->addSql('ALTER TABLE neos_flow_security_account RENAME INDEX flow_identity_neos_flow_security_account TO flow_identity_typo3_flow_security_account');
+        } else {
+            $this->addSql('DROP INDEX flow_identity_neos_flow_security_account ON neos_flow_security_account');
+            $this->addSql('CREATE UNIQUE INDEX flow_identity_typo3_flow_security_account ON neos_flow_security_account (accountidentifier, authenticationprovidername)');
+        }
     }
 }
